Removed redundant test.
